Move tests around to match project file layout
